fix: render fails if @aws-cdk/core is not a dependency
if a module doesn't have @aws-cdk/core as a dependency, the renderer fails because the renderer looks up `@aws-cdk/core.Construct`. To address this, if @aws-cdk/core is not in the closure, we just install it manually so the renderer is happy.

An example of such a module is `cdk-constants` which is basically just a set of constants.